+++ Source/SelfService/Web/microservice/microserviceView/configuration/restartInfoBox.tsx
